refactor: sepparate level ui into it's own class
